refactor(kafka-tools): added a new shared folder
