+++ apps/kafka-tools/project.json
@@ -20,8 +20,8 @@
                 ],
                 "styles": [
                     "apps/kafka-tools/src/styles.scss",
-                    "apps/kafka-tools/src/_styles/css-reset.scss",
-                    "apps/kafka-tools/src/_styles/theme.scss"
+                    "apps/kafka-tools/src/shared/_styles/css-reset.scss",
+                    "apps/kafka-tools/src/shared/_styles/theme.scss"
                 ],
                 "scripts": []
             },
