--- apps/kafka-tools/src/_styles/themes/git-hub/git-hub.theme.scss
@@ -1,6 +1,0 @@
-@import '~/src/_styles/functions/create-theme.function.scss';
-@import '~/src/_styles/themes/git-hub/git-hub.pallette.scss';
-
-$git-hub-theme: create-theme($git-hub-pallette);
-
-$git-hub-theme-export: (git-hub, $git-hub-theme, $git-hub-pallette);
