--- apps/kafka-tools/src/_styles/themes/light/light.theme.scss
@@ -1,6 +1,0 @@
-@import '~/src/_styles/functions/create-theme.function.scss';
-@import '~/src/_styles/themes/light/light.pallette.scss';
-
-$light-theme: create-theme($light-pallette);
-
-$light-theme-export: (light, $light-theme, $light-pallette);
