--- apps/kafka-tools/src/_styles/themes/pallettes.scss
@@ -1,11 +1,0 @@
-@import '~/src/_styles/themes/dark/dark.pallette.scss';
-@import '~/src/_styles/themes/light/light.pallette.scss';
-@import '~/src/_styles/themes/git-hub/git-hub.pallette.scss';
-@import '~/src/_styles/themes/pop-n-lock/pop-n-lock.pallette.scss';
-
-$pallettes: (
-    (dark, $dark-pallette),
-    (light, $light-pallette),
-    (git-hub, $git-hub-pallette),
-    (pop-n-lock, $pop-n-lock-pallette)
-);
