+++ apps/kafka-tools/src/app/app.component.scss
@@ -1,4 +1,4 @@
-@import '~/src/_styles/theme.scss';
+@import '~/src/shared/_styles/theme.scss';
 
 :host .content {
     height: 100%;
