+++ apps/kafka-tools/src/shared/_interfaces/kafka-message.model.ts
