+++ apps/kafka-tools/src/shared/_styles/functions/create-theme-vars.function.scss
