+++ apps/kafka-tools/src/shared/_styles/functions/get-darkens.function.scss
