+++ apps/kafka-tools/src/shared/_styles/functions/get-lightens.function.scss
