+++ apps/kafka-tools/src/shared/_styles/themes/dark/dark.pallette.scss
