+++ apps/kafka-tools/src/shared/_styles/themes/dark/dark.theme.scss
@@ -1,0 +1,6 @@
+@import '~/src/shared/_styles/functions/create-theme.function.scss';
+@import '~/src/shared/_styles/themes/dark/dark.pallette.scss';
+
+$dark-theme: create-theme($dark-pallette);
+
+$dark-theme-export: (dark, $dark-theme, $dark-pallette);
