+++ apps/kafka-tools/src/shared/_styles/themes/git-hub/git-hub.theme.scss
@@ -1,0 +1,6 @@
+@import '~/src/shared/_styles/functions/create-theme.function.scss';
+@import '~/src/shared/_styles/themes/git-hub/git-hub.pallette.scss';
+
+$git-hub-theme: create-theme($git-hub-pallette);
+
+$git-hub-theme-export: (git-hub, $git-hub-theme, $git-hub-pallette);
