+++ apps/kafka-tools/src/shared/_styles/themes/light/light.pallette.scss
