+++ apps/kafka-tools/src/shared/_styles/themes/pop-n-lock/pop-n-lock.theme.scss
@@ -1,0 +1,6 @@
+@import '~/src/shared/_styles/functions/create-theme.function.scss';
+@import '~/src/shared/_styles/themes/pop-n-lock/pop-n-lock.pallette.scss';
+
+$pop-n-lock-theme: create-theme($pop-n-lock-pallette);
+
+$pop-n-lock-theme-export: (pop-n-lock, $pop-n-lock-theme, $pop-n-lock-pallette);
